+++ src/euphorie/client/survey.py
@@ -167,7 +167,7 @@ class Identification(grok.View):
     grok.template("identification")
     grok.name("index_html")
 
-    question_filter = model.RISK_OR_MODULE_WITH_DESCRIPTION_FILTER
+    question_filter = None
 
     def update(self):
         if redirectOnSurveyUpdate(self.request):
@@ -200,10 +200,8 @@ class Evaluation(grok.View):
     grok.template("evaluation")
     grok.name("index_html")
 
-    question_filter = sql.and_(
-            model.RISK_OR_MODULE_WITH_DESCRIPTION_FILTER,
-            sql.or_(model.MODULE_WITH_RISK_NO_TOP5_NO_POLICY_FILTER,
-                model.RISK_PRESENT_NO_TOP5_NO_POLICY_FILTER))
+    question_filter = sql.or_(model.MODULE_WITH_RISK_NO_TOP5_NO_POLICY_FILTER,
+                              model.RISK_PRESENT_NO_TOP5_NO_POLICY_FILTER)
 
     def update(self):
         if redirectOnSurveyUpdate(self.request):
@@ -233,10 +231,8 @@ class ActionPlan(grok.View):
     grok.template("actionplan")
     grok.name("index_html")
 
-    question_filter = sql.and_(
-            model.RISK_OR_MODULE_WITH_DESCRIPTION_FILTER,
-            sql.or_(model.MODULE_WITH_RISK_OR_TOP5_FILTER,
-                              model.RISK_PRESENT_OR_TOP5_FILTER))
+    question_filter = sql.or_(model.MODULE_WITH_RISK_OR_TOP5_FILTER,
+                              model.RISK_PRESENT_OR_TOP5_FILTER)
 
     def update(self):
         if redirectOnSurveyUpdate(self.request):